update functions js signature
