+++ djgentelella/firmador_digital/consumers/pdf_render.py
@@ -1,0 +1,8 @@
+from rest_framework.renderers import BaseRenderer
+
+class PDFRenderer(BaseRenderer):
+    media_type = 'application/pdf'
+    format = 'pdf'
+
+    def render(self, data, accepted_media_type=None, renderer_context=None):
+        return data
